move songlist to session collection (#43)
* move song db functions to session
- implement GetSongByID
- implement AddSong
- implement RemoveSong

* - implement ListSongs
- sort songs in AddSong

* fix some comments

* Format error messages (#44)

* Format error messages

* Fix info string

* implement voteUP

* implement voteDown

* removed unused method

* fixed missing error

* wrap errors in Vote methods

* removed unused variables

* dont update user score if votes on own song

* fix tests

* regenerate mocks

* add tests for remove song

* impl userhandler tests

* move song db functions to session
- implement GetSongByID
- implement AddSong
- implement RemoveSong

* - implement ListSongs
- sort songs in AddSong

* fix some comments

* implement voteUP

* implement voteDown

* removed unused method

* fixed missing error

* wrap errors in Vote methods

* removed unused variables

* dont update user score if votes on own song

* fix tests

* regenerate mocks

* add tests for remove song

* impl userhandler tests

* implement tests for voteUp

+++ user/model.go
@@ -8,22 +8,22 @@ import (
 )
 
 type Model struct {
-	ID                string  `json:"id" bson:"_id"`
-	Username          string  `json:"username" bson:"username"`
-	Secret            string  `json:"secret" bson:"secret"`
-	SessionID         string  `json:"session_id" bson:"session_id"`
-	IsAdmin           bool    `json:"is_admin" bson:"is_admin"`
-	Score             float64 `json:"score" bson:"score"`
-	SpotifyAuthorized bool    `json:"spotify_authorized" bson:"spotify_authorized"`
+	ID                string `json:"id" bson:"_id"`
+	Username          string `json:"username" bson:"username"`
+	Secret            string `json:"secret" bson:"secret"`
+	SessionID         string `json:"session_id" bson:"session_id"`
+	IsAdmin           bool   `json:"is_admin" bson:"is_admin"`
+	Score             int    `json:"score" bson:"score"`
+	SpotifyAuthorized bool   `json:"spotify_authorized" bson:"spotify_authorized"`
 
 	AuthToken *oauth2.Token `json:"auth_token" bson:"auth_token"`
 	AuthState string        `json:"auth_state" bson:"auth_state"`
 }
 
 type ListElement struct {
-	Username string  `json:"username" bson:"username"`
-	IsAdmin  bool    `json:"is_admin" bson:"is_admin"`
-	Score    float64 `json:"score" bson:"score"`
+	Username string `json:"username" bson:"username"`
+	IsAdmin  bool   `json:"is_admin" bson:"is_admin"`
+	Score    int    `json:"score" bson:"score"`
 }
 
 func GenerateUserID(username, userID string) string {
